+++ python/ray/serve/endpoint_state.py
@@ -4,7 +4,7 @@ from ray import cloudpickle
 from ray.serve.common import (BackendTag, EndpointInfo, EndpointTag,
                               TrafficPolicy)
 from ray.serve.long_poll import LongPollNamespace
-from ray.serve.kv_store import RayInternalKVStore
+from ray.serve.storage.kv_store import RayInternalKVStore
 from ray.serve.long_poll import LongPollHost
 
 CHECKPOINT_KEY = "serve-endpoint-state-checkpoint"
